feat: move WebConfigurator to configuration mode
